Set workflow permissions
Since 2023, it has been possible to limit the permissions granted to
a workflow. The default of all permissions is now being flagged as
a security issue, so add some minimal permissions.

It isn't really clear from the docs which of these are actually
required, as the only detailed docs are at the web API level, but
libraries for working with actions are generally in Python or NodeJS
and the libraries have not been updated to state which permissions
they require.