Yieri pH meter: add pH-W3988
RC Yago pH-W3988 seems to have the same sensors in the same dps with same
scaling, and also the cosmetic design matches Yieryi branded devices, so
likely a compatible model from the same underlying manufacturer.

There are a couple of extra sensors, and a set of availability flags.
Curiously, the availability flags are missing the extra sensors.

- CF seems to be a duplicate EC reading on a different (non-standard) unit scale
- RH = Relative Humidity, a bit strange for water readings, so maybe not really included? Added as optional, hidden if missing.

As well as the sensors, this also defines minimum and maximum alarm thresholds for each value. Use the flags to make these hidden if not supported.

Issue #2976