Lock: add lock_state for read-only reporting of lock state
It seems quite common for locks to have a read-only "lock_motor_state"
dp that reports the status of the lock. This is separate from any dp
for performing locking and unlocking if there is one, so having a
read-only lock dp does not cover all the possibilities.

Based on proposed change in PR #3063.