PG103/107: restore binary sensors for battery and power
Although previous comment on #1059 suggested that these act more as events,
and do not correctly reset when the power or battery is restored, based
on new information on #3272 they appear to be indicating that the alarm is
sounding for these events, and will correctly reset when the alarm is stopped.

Therefore they are useful as state sensors to check if the alarm is sounding.