Cloud config_flow: avoid using hub product id when adding sub device
When adding a subdevice with cloud config flow, we do a switch part
way through the process to using the hub's details with cid from the
sub device, since local connection details are mostly from the hub.

But this was also switching the product id, at a point before it is
used for config matching, which is incorrect. A related problem was
that during the network scan, the local productKey for the hub would
be reported, and included along with the cloud product id for matching
later.

Ensure that the product id from the sub device goes all the way
through to the matching code, and do not use the local productKey
for subdevices, since we do not see those directly on the local
network.

Issue #3280