Kettles: avoid using non-standard operation modes
HA documentation for water_heater seems to have been updated since it
was implemented, to clarify that integrations must only use
pre-defined operation modes.

So using operation_mode for different kettle modes is not correct.

Instead, move the translations to a kettle_mode key under select
entity, and add such an entity to kettles that were previously abusing
operation_mode for this, and use operation_mode only as a switch
between "off" and "electric".

- Korex AX-WF306N disentangle the two temperature settings, and
instead allow temperature_unit to take effect. Since the temperatures
are no longer entangled, and the keep warm switch is separated out,
only have the "keep warm" temperature as a separate select entity.

The exact operation of this device is still a bit confusing to me,
especially with the number of write-only dps related to temperature
settings that are currently being ignored by the config.

PR #3301