feat(feyree_ev_portable_charger): expand to cover higher capacity models
According to discussion #3373 the 22kW feyree charger is almost compatible
except for target range. Full details for the same product id is in
issue #3355, so expand this config to cover the full dps of this series.

Obviously some of these are only relevant to the higher power models,
and 3-phase, so some are hidden by default to keep things simple for the
single-phase installations, which are perhaps more common in domestic
use cases.