Add support for Howeall Register Booster Fan (#3444)
* Adding Howeall register booster fan

* Make this be an actual fan entity (vs separate power and speed)

* feat(howeall_register_booster_fan): use climate for control

Since this has heat and cool modes with temperature trigger levels, it
makes sense to model as a climate entity. Though the large number of speeds
is more conveniently controlled with a fan entity, so it cannot be completely
replaced.

Add some info from issue #3443, the 4 missing "sensors" appear to be for
setting minimum and maximum speeds for the heat and cool modes, though
that is not easily implemented here. There also appears to be another
mode SLEEP, which was missing from the original submission, and can only
really be modelled as a preset_mode, or by abusing the unused hvac_modes to
fake it as one of them. The temperature range also says 32-112 there, but
that seems too low and high to be useful as triggers, so left at 60-90.

PR #3443

---------

Co-authored-by: Jason Rumney <make-all@users.noreply.github.com>
Co-authored-by: Jason Rumney <jasonrumney@gmail.com>