feat(fersk_vind_2_climate): add compatible Pro Breeze model
Fill in the unknowns based on the info for the Pro Breeze 12000 BTU model.
Due to unusual similarities in the modes, most likely these are correct for
both models.
- reduced max temperature to 30C/86F to not go out of range on the Pro Breeze.
Since there was no info on the original PR, it may also be correct for that,
and the 32/90 limit was just a copy and paste from another model.

Issue #3478