fix(goldair_portable_airconditioner): use swing_mode when only one
When a device has only horizontal swing mode (seems usual for portable
air conditioners), redirect it to swing_mode, as HA expects swing_mode
to be supported primarily, so does not really support devices with only
swing_horizontal_mode.

This reverts an earlier change that swapped vertical and horizontal swing
modes, which was probably incorrect, as the documentation for goldair
clearly says dp 15 should be vertical swing.

Issue #3496