feat: use time entity for 24h timers
Time provides a better UI for entering times than Number, but is
limited to hh:mm input where the maximum time is 23:59. This is good
for 24h timers in addition to wall clock time (alarms etc).

Timers that are just h are left as numbers, to avoid confusing the
user into thinking they can set minutes.

Many of these timers have second precision, which is not exposeed
in the UI, but should work when set via an action.

Issue #3499