feat: add support for MeacoCool MC Series Pro air conditioners
Although user reports that swing is not supported on either the
C or CH 10k btu model, the dp is there and there is no info to
determine that, while in manuals for the non-Pro models swing is
mentioned, so leave it in for other potential models. Worst case
it is there but does nothing.

Issue: #3393