fix(time): timedelta does not like None arguments
When the device is not sending optional dps, components can be set to None
The last change attempted to detect this condition and return None rather
than trying to construct a timedelta, but since we initialise the unused
components to 0, the detection failed. It could also fail if there are two
components of time and only one is missing, so change the logic to cover all
cases.

Discussion #3588