fix(hy_combo_heatpump): improve detection
Based on user experience, many of the dps can be missing in the
initial data returned by the device. Mark these as optional, and
handle the case of them missing (some like "mode" are quite fundamental
to the device working correctly, so hopefully this is only a transient
problem and their real values are returned soon after starting.

Also added product id that was previously missing.

Issue #2510