fix(feyree ev chargers): allow current to be set down to 6A
Based on multiple reports for multiple chargers, it is common for EV
chargers to be settable down to 6A, even though they are documented
with minimum of 8A. This seems to be the case at least for 16A and 32A
models, and is especially requested by users who use solar panels with
them to maximise use of solar power without using grid power for
charging.

Issue #3697 (requested for feyree_ev_portable_charger, but assume the
same is true for feyree_ev_charger also).