fix (device_config): allow range in mapping to override
Previous code would override default range with one in a conditional mapping,
but not in a simple mapping. I guess there was not much point in that
before now, but with a value_redirect it does make sense to be able
to change the range of a dp in a simple mapping.

Issue #3779