fix (breville_airrounderplus_heater): split fan and heat
Due to the complex interaction of dps 1, 4, and 5, it is difficult to
combine all the features into a climate entity. When fan is active,
dp 4 goes to heat_low, and when heat is active, dp 5 goes to SleepWind,
but setting those dps to those values does not automatically start
the other mode.

Another approach that reduces the complexity is to only support heat
mode in the climate entity, and use a separate fan entity for the fan modes

With this configuration, turning the climate on will enable the heating mode
and automatically show the fan as off. Turning on the fan will enable fan
mode and automatically show the climate entity as off. Turning either off
will turn off the switch for the whole device.

Due to the limitation of having to ensure a valid mode is set when turning
on, the middle setting will be automatically activated

Another behaviour which may be surprising (but also useful once it is
understood) is that the climate toggle action will toggle the switch
without changing dp 4 or 5, which should turn the device on with the
last settings. This means that the fan might be turned on instead of
the climate entity by the climate toggle action.

Issue #3582