fix (devices): not not use unicode micro sign, use greek mu
Home Assistant standardised on using Greek small letter mu for micro
units a couple of versions back. Some configs were using the micro
sign, which is apparently deprecated by Unicode.

For μg/m³, use the abbreviation `ugm3`, which is handled in code using
the HA constant for the unit.

For others (μS/cm for conductivity) which do not have a native HA
unit, make sure the correct Unicode character is used.