fix (feyree_ev_portable_charger): drop RFID and make some optional
Recent changes to support more models from Feyree without making new
configs every time have broken compatibility with the product this
config was originally made for.
The boolean RFID switch on other models clashes with string "PE" dp that
was unsupported by the original config. Some of the sensors that were added
later are actually listed in the Query Things Data Model for the original
device, but were not appearing in its logs, so need to be optional.

Issue #3860