feat (ledvance_smart?plabl100x25b): add compatible Annaror light
These devices both have separate white and rgb lights, controlled through
dp 51 alongside standard light dp layout which is not used except as a master
switch. The Annaror does not have enough documentation to confirm, but
I suspect the format for dp 51 is the same, and the same config can be used.

Issue #3883