fix (device_config): use pending changes when setting multiple dps
When a dps is reused multiple times with different masks, and those
attributes get set together in a single function, the different values
were overwriting each other.

In this case, allow the pending changes that have already been
calculated but not actioned yet to be passed in so they can be used
instead of the current value from the device in further
recalculations.

A more general solution to the suggested change for time entity in
PR #3902