fix (device_config): Keep dps_map and pending_map separate
Tests appear to be failing because the dps_map contents are being carried over.
It may be returning a map that was passed in then modified and gets combined
with the same map that is causing this weird behaviour, so return a fresh
map every time to avoid this.

Refining PR #3902 alternative solution