Synchronize accesses to sub-devices (#3929)
When multiple sub-devices are under the same parent device, we need to make sure to access only one sub-device at a time,
because the underlying tinytuya device for the parent cannot handle multiple acceses at the same time. This bug often
manifests as "No local key for device!" in the tinytuya logs.

This commit adds a lock that is associated with the parent device, and acquires the lock whenever we access a sub-device or
the parent device itself.

Issue #2289