feat (lock): allow locking and unlocking with a code
Tuya locks following the standard BLE lock template use an 8 digit
code for unlocking operations (and sometimes for locking as well).
This is encoded as base64 on dp 61 remote_no_dp_key, as:
  Byte 1: Action (0x00 = Lock, 0x01 = Unlock)
  Byte 2-3: User (0x0001 - 0x0064)
  Byte 4-11: Code (ASCII)
  Byte 12-13: Source (0x0000 = Unknown, 0x0001 = App, 0x0002 = Voice)
  Byte 14: Admin flag (0x00 = normal user, 0x01 = admin user)

The above format is supported as code_unlock in the config, using
hardcoded values for user=0x0001, Source=0x0000, Admin=0x00, but the
8 digit code needs to be extracted from elsewhere by the user and
passed in to the unlock method.

Related to Issue #1921 (slightly different, as the request there was
for setting temporary and offline passwords for manual use on the
device keypad, while this is using codes for remote unlocking).