device (subdevice sync): be fault tolerant getting cached lock
At startup I am seeing failures in the log due to missing tuyadevicelock
I can't immediately spot where, but there must be a path where this is
not getting initialised after a restart. Use get() instead of directly
accessing, with a default of a new Lock.

Also clean up the sleep calls so the lock does not need to be released
mid loop.

PR #3929