fix (device_config): redirect first when setting values.
If the current dps_val has a redirection, then use that first.
Mostly this is used for null mappings to redirect to an alternate dp,
which we want to work in both directions.
Previously I think this had to be in a condition block to work, but
it is more obvious to do it directly in the mapping in the null handling case.

Issue #3987