fix (immax_neolite_weatherstation): match without outdoor unit
Config was originally made with the extra remote sensors optional, but the
assumption that data from the main outdoor unit is always available.
Issue #4041 seems to be missing all the outdoor sensor info, so it seems
some users will try to set up the indoor unit only first.

Changed one of the duplicate product details for immax to the matched Emax
product. It isn't clear which is the OEM, likely these are a rebadging of
the same hardware, though the previous report noting a match for this
product id may have been for Emax branded hardware too.