dev (utils): add an entities utility
Takes one argument, a entity config id made up of entity type
followed optionally by an underscore and name, translation_key,
or class (in that order)

Outputs the matching entities, with filename and line info compatible
with grep and compiler output so it can be easily parsed by IDEs.

Mainly to assist with #1708, but probably useful in other cases too.