cleanup! (devices): standardise various mode selects.
Breaking change, as entity ids and option values are changed. As there
is not a simple 1-to-1 mapping for the new ids, and multiple source
ids and independently multiple target ids, migration is difficult and
would anyway not take care of value changes in existing automations.

Uses the translations from previous commit.

Issue #1708