feat (etop_hy_thermostat): add compatible BVF 801
Compatible dps layout, but is missing the window open detection switch
though it does list "window_opened" as one of the action states.

Adapt hvac_action to handle the expanded states, and add a binary_sensor
for the window_opened state.

Disable the window open detection switch when unavailable.

Issue #4079