fix (ipengen_goe3116a_evcharger): shuffle energy dps
According to user feedback, dp 1 remains unknown (no data received), and
the data on dp 17 is the actual energy consumption reading from the device.

This is contrary to the initial assumption that because dp 17 is read-write
it is for setting the fixed charge energy amount (for charge_energy mode),
and dp 1 being labelled total_forward_energy in the data model from Tuya and
read-only was the energy consumption.

Looking back on issue #3847, the original reporter had posted a config with
dp 1 and 25 commented out, backing up the evidence that these dp are not
actually used by the device.

Issue #4057