feat (climate): allow precision to be specified for current temp
When there is a unit mismatch, it is useful to be able to scale
into the expected range using target_range. But this results in fractional
temperatures, which are unexpected. Allow precision to be specified in
the current_temperature dp, and use it to round the value after scaling.

This is different to the use of precision in sensors to inform the UI of a
suggested precision without actually rounding, but it makes sense in this
context.

For PR #4051