fix (ilife_a30_pro_vacuum): invert the activate signal
Since the device defines this as "pause", it is inverted compared to an
actual activate dp found on some other vacuums.

This is a simpler change than the approaches used in kogan_lx8_vacuum
(combining pause with command), and kabum_smart700_v2023_vacuum (combining
switch_go with pause into a single activate dp) if it works.

PR #4109