feat (sensor): suggest actual sensor precision if not defined
In HA 2025.6 release, a change went in that redefined the default precision
to be based on sensor class rather than using all precision that the sensor
provides. This was noticed on a battery voltage sensor where the default
precision for voltage class is 0, which may be appropriate for mains voltages,
but not for a 3.6V battery voltage.
Rather than review all 1300+ configs for sensors that need to have more
precision than the new HA defaults provide, provide a default precision based
on native_precision rather than None.

PR #4119