fix (migration): remove rather than update entity entries with invalid id
Since entity entries will be automatically created when the entities
are created, the invalid entity entries have already gone dormant and
been replaced by new entries with the correct id. The earlier attempt
to fix the unique id was therefore unsuccessful, as it duplicated the
already recreated entry.

Instead, remove the invalid entries.

Bump the minor version so this gets run regardless of whether the
migration to 3.11 already ran successfully or not.

Issue #4130