Update auchsiag_fireplace_750w_1500w.yaml
I have implemented the translation keys for the common entities (timer, power, time remaining) as requested.

However, regarding the light entities: I tested it, and I must revert to select entities for the Flame and Ember control.

Technical reason: This device shares the same DP (id: 1) for the main power and the "light" power. It is impossible to turn off the flame/ember light without turning off the heater itself. Using a light entity in Home Assistant is misleading here, because clicking the "turn off" button on the light entity shuts down the entire fireplace (including heat), which is bad UX.

Since the colors are fixed presets and cannot be toggled independently of the heating unit, select entities are the most accurate representation of the hardware capabilities.