Create auchsiag_fireplace_750w_1500w.yaml (#4172)
* Create auchsiag_fireplace_750w_1500w.yaml

Add support for AUCHSIAG Fireplace

* Update auchsiag_fireplace_750w_1500w.yaml

* Update auchsiag_fireplace_750w_1500w.yaml

Fix YAML indentation for sensor

* Update auchsiag_fireplace_750w_1500w.yaml

* Update auchsiag_fireplace_750w_1500w.yaml

Rework: convert to lights and use translation keys

* Update auchsiag_fireplace_750w_1500w.yaml

I have implemented the translation keys for the common entities (timer, power, time remaining) as requested.

However, regarding the light entities: I tested it, and I must revert to select entities for the Flame and Ember control.

Technical reason: This device shares the same DP (id: 1) for the main power and the "light" power. It is impossible to turn off the flame/ember light without turning off the heater itself. Using a light entity in Home Assistant is misleading here, because clicking the "turn off" button on the light entity shuts down the entire fireplace (including heat), which is bad UX.

Since the colors are fixed presets and cannot be toggled independently of the heating unit, select entities are the most accurate representation of the hardware capabilities.

* Update auchsiag_fireplace_750w_1500w.yaml

I have found the correct Product ID (cokyjkkw8gqwteww) and updated the configuration accordingly

* feat (auchsiag_fireplace) refactor

- rename to follow file naming convention
- 750 & 1500 is not really a model name, more a description, so drop it
- since the device cannot be on without the lighting, use the embers as main switch, and merge brightness and ember colors into a standard light entity.
- leave flame as a select, as some color options are combos.
- make other entities availability dependent on the embers switch to avoid confusion
- add quotes around icons, as they contain embedded colons which confuse the yaml parser
- remove icons of translated entities
- add class for time remaining sensor. Guess the units as min, which is common for hour select timers.

PR #4172

* whitespace cleanup

---------

Co-authored-by: Jason Rumney <make-all@users.noreply.github.com>