Create inkbird_ivc001w_fan.yaml
Device is unusual in that it does not follow the normal design of either a thermostat or HVAC device and is much more similar to a fan. 

Also does a very non standard base64 encoded string to show current sensor readings that needs a set of template sensors to expose. 

Would be amazing to have the ability to do "virtual" DP's and base64 decoding easily. 

I left the comments in for the bits that really need explaining as the device is weird - most people using this in HA will purely be looking to easily control the fan speed rather than using all of the capabilities of the device.