Add configuration for Costco Solar bird feeder camera
Added configuration for Solar smart bird feeder camera with various entities including control of the squirrel deterrence siren.

Based on pinelake_birdfeeder.yaml with minor additions for this slightly newer version.

dp_id 127 is event detected but doesn't seem to ever be anything but
motion, but could be human according to the api but its read-only it seems.

dp_id 233 is ipc debug returns some big long string of info but doesn't
display well in HA.