fix (arcelik_ahphmm_comboheatpump): handle off edge cases
When in hot_water mode, the water_heater needs to be turned off by
the power switch. The redirect in this case hides the available modes,
since the other modes are filtered out to avoid affecting the climate mode.
Add some dummy conditions to make the options appear in the list.

In climate mode, "off" was only available when hot water was on, similarly
because "off" in that case requires the use of the power switch. Make
that available with a redirection also.

The reverse cases of turning the device on from power off mode needs to be
done from the main power switch, as the other entities are disabled to
avoid needing to send two dps at once to switch modes out of "off".

PR #4200